Finished organizing the components folder. Added a new MessageItem component representing a single message item.
renamed previous MessageItem component to MessagesList to accurately reflect it's function.

added scss typing to all style tags for consistency.
