+++ src/views/ChatLayout.vue
@@ -6,9 +6,9 @@ import { ChevronDown } from 'lucide-vue-next';
 import { determineModelDisplayName, handleDoubleClick } from '@/libs/utils/general-utils';
 import { handleExportConversations } from '@/libs/conversation-management/conversations-management';
 import { uploadFileContentsToConversation, uploadFile, imageInputChanged } from '@/libs/file-processing/file-processing';
-import messageItem from '@/components/MessageItem.vue';
-import chatInput from '@/components/ChatInput.vue';
-import chatHeader from '@/components/ChatHeader.vue';
+import messageItem from '@/components/controls/MessagesList.vue';
+import chatInput from '@/components/layout/ChatInput.vue';
+import chatHeader from '@/components/layout/ChatHeader.vue';
 import settingsDialog from '@/components/dialogs/SettingsDialog.vue';
 import conversationsDialog from '@/components/dialogs/ConversationsDialog.vue';
 import StoredFilesList from '@/components/dialogs/StoredFilesDialog.vue';
